+++ src/main/java/joey/command/IndexedCommand.java
@@ -1,6 +1,6 @@
-package commands;
+package joey.command;
 
-import exceptions.CommandFormatException;
+import joey.exception.CommandFormatException;
 
 public abstract class IndexedCommand implements Command {
     protected final int taskIndex;
@@ -11,18 +11,18 @@ public abstract class IndexedCommand implements Command {
 
     protected static int parseTaskIndex(String commandName, String userInput) throws CommandFormatException {
         if (userInput.trim().length() <= commandName.length()) {
-            throw new CommandFormatException("The task index cannot be empty.");
+            throw new CommandFormatException("The joey.task index cannot be empty.");
         }
 
         String[] parts = userInput.split(" ");
         if (parts.length < 2) {
-            throw new CommandFormatException("Specify a task number after '" + commandName + "'.");
+            throw new CommandFormatException("Specify a joey.task number after '" + commandName + "'.");
         }
 
         try {
             return Integer.parseInt(parts[1]) - 1; // 0-based index
         } catch (NumberFormatException e) {
-            throw new CommandFormatException("'" + parts[1] + "' is not a valid task number.");
+            throw new CommandFormatException("'" + parts[1] + "' is not a valid joey.task number.");
         }
     }
 }
