+++ src/main/java/joey/command/UnmarkCommand.java
@@ -1,24 +1,24 @@
-package commands;
+package joey.command;
 
-import exceptions.CommandFormatException;
-import exceptions.TaskIndexOutOfBoundsException;
-import storage.Storage;
-import task.TaskList;
-import ui.Ui;
+import joey.exception.CommandFormatException;
+import joey.exception.TaskIndexOutOfBoundsException;
+import joey.storage.Storage;
+import joey.task.TaskList;
+import joey.ui.Ui;
 
 import java.io.IOException;
 
-public class MarkCommand extends IndexedCommand {
-    public MarkCommand(String userInput) throws CommandFormatException {
-        super("mark", userInput);
+public class UnmarkCommand extends IndexedCommand {
+    public UnmarkCommand(String userInput) throws CommandFormatException {
+        super("unmark", userInput);
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws CommandFormatException {
         try {
-            tasks.markTask(taskIndex);
+            tasks.unMarkTask(taskIndex);
             storage.save(tasks);
-            ui.showMarkedTask(tasks.getTask(taskIndex), taskIndex + 1); // 1-based index for UI
+            ui.showUnmarkedTask(tasks.getTask(taskIndex), taskIndex + 1);
         } catch (TaskIndexOutOfBoundsException e) {
             throw new CommandFormatException("Task " + (taskIndex + 1) + " does not exist.");
         } catch (IOException e) {
