+++ src/main/java/joey/exception/CommandFormatException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package joey.exception;
 
 public class CommandFormatException extends Exception {
     public CommandFormatException(String message) {
