+++ src/main/java/joey/exception/TaskIndexOutOfBoundsException.java
@@ -1,7 +1,7 @@
-package exceptions;
+package joey.exception;
 
 public class TaskIndexOutOfBoundsException extends Exception {
-    private static final String ERROR_MESSAGE = "task.Task Index out of range";
+    private static final String ERROR_MESSAGE = "joey.task.Task Index out of range";
 
     public TaskIndexOutOfBoundsException() {
         super(ERROR_MESSAGE);
