+++ src/main/java/joey/storage/Storage.java
@@ -1,6 +1,6 @@
-package storage;
+package joey.storage;
 
-import task.*;
+import joey.task.*;
 
 import java.io.*;
 import java.nio.file.Files;
