+++ src/main/java/joey/task/Task.java
@@ -1,4 +1,4 @@
-package task;
+package joey.task;
 
 public abstract class Task {
     private boolean isDone;
@@ -25,7 +25,7 @@ public abstract class Task {
         return this.description;
     }
 
-    // For storage purposes
+    // For joey.storage purposes
     public abstract String getStorageFormat();
     public static Task createFromStorage(String data) {
         // Provide implementation in a subclass
