+++ src/main/java/joey/task/Todo.java
@@ -1,6 +1,6 @@
-package task;
+package joey.task;
 
-import enums.TaskType;
+import joey.enums.TaskType;
 
 public class Todo extends Task {
     private final TaskType type;
